Add battle cards before war cards recursively
Gets the fifth and sixth test case to pass, leaving one test case
(number-7; the long game one) that still fails (it ends unexpectedly in
a draw).

Also the "one game, one battle" test case (number-5) has three wars, so
it may be passing by luck somehow.

It would have been nice if the problem statement included a short
example since the wording is ambiguous on how to order the winning
cards for a war or chain of wars.